requirements.txt
asgiref==3.8.1
Django==5.0.4
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.0
iniconfig==2.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
pytest==8.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
